+++ core/market/decentralized/src/testing/mock_node.rs
@@ -100,10 +100,11 @@ impl MarketsNetwork {
         // level 2 is <core::future::from_generator::GenFuture<T> as
         // core::future::future::Future>::poll::XXX> (async)
         // We want to know the caller.
-        let mut bn = crate::testing::generate_backtraced_name(Some(3));
-        // Special case for mac&windows
+        let mut bn = crate::testing::backtrace_util::generate_backtraced_name(Some(3));
+        // Special case for mac&windows. Tests are run in adifferent way on those systems and we
+        // have to dive one less level down the stack.
         if !bn.starts_with("test_") {
-            bn = crate::testing::generate_backtraced_name(Some(2));
+            bn = crate::testing::backtrace_util::generate_backtraced_name(Some(2));
         }
         let tn = test_name.unwrap_or(bn.as_str());
         log::info!("Intializing MarketsNetwork. tn={}", tn);
